Add initial, crude database import script
This uses [roo][1] to parse the EFile Excel spreadsheet into a SQLite
database. Currently, as I continue to grapple with the data, it only
imports the first tab "A-Contributions". This is Schedule A of [CA Form
460 (page 5)][2].

I made a design decision to branch into multiple tables by type of
contributor (and also recipient, though currently there can't be any
other type than 'Committee'). The contributor branching is based on the
'Entity_Cd' column, the values of which are described in the legend at
the bottom of Form 460 page 5.

For now, I decided to commit the sqlite database so anyone fetching this
can see the output without setting up the Ruby environment.

(This is still a WIP, I will be restructuring the file organization and
addressing some of the inline TODOs in future commits.)

[1]: https://github.com/hmcgowan/roo
[2]: http://www.fppc.ca.gov/forms/rev201208/460.pdf