Add hacky way to deal with Socrata data
This pulls the Socrata data every 24 hours and caches it in memory. But
only in the `production` RACK_ENV - otherwise the data_local.csv is
used, which actually has a slightly different format than the Socrata
data (hence the conditional in app.js). This will be rectified with a
future commit.