Add basic backbone views
So, this converts the Mayoral table into something that is closer to how
we want to end up doing it. On page load, it loads GET /api/candidates
and then renders a Backbone view with the data.

All the fun happens in the `assets/js/backbone.js` file, so that is a
good place to look to make sense of this whole thing.

Next up is modifying the existing charts to use this data instead of
requesting the CSV directly.