Add sass-enhance library
I may just be in a rut from my day job, but we open-sourced this tiny
bit of Sass which is super handy. You can read more about it here:

https://github.com/causes/sass-enhance

tl;dr: to apply styling that happens at breakpoints equal to or larger
than a given breakpoint, put styling inside

    @include enhance(medium)

where "medium" is a bootstrap breakpoint size.

To apply styling for breakpoints equal to or smaller than a given
breakpoint, use

    @include degrade(medium)