Also include candidates without data in /api/candidates
This is relevant to #133 because we're going to add all candidates to
the table, regardless of whether they have funding. But if they don't
have funding, we can't get them from the database! So, this adds a bit
more logic to the /api/candidates endpoint to look at the
Party::CANDIDATE_INFO constant to see which keys are not in the
Party::MAYORAL_CANDIDATE_IDS array and instantiate new Party::Committee
objects with that data.