Parse data in chronological order (#323)
Since we overwrite Summary data as we parse data (so as to only maintain
the most recent summary figures for each candidate), this produces
nondeterministic behavior for candidates with multiple summaries.

Adding an '$order' to the query when we retrieve the data from Socrata
should provide reproducible ordering of data.

Parsing the data in this order results in errors for the 14
contributions with no Tran_NamL value. I added in a simple message in
that case.