Refactor Fetching structure
The main idea is to separate the downloading of the data from the
processing of it, under the assumption that the Cal-Access data will be
similar enough that it can be massaged into something looking like it
came from Socrata without too much hassle.

So, there are Downloaders (of which there will be two - Socrata and
Cal-Access) and Fetchers (a bad name, but they process the downloaded
data).

Also, this adds a preliminary amount of test coverage for the most basic
processing conditions, and hopefully this will give us all some
motivation to be more conscious about writing tests when we find bugs. I
ripped out FactoryGirl and replaced it with parsing raw JSON dumps,
which seems easier to understand.