Remove Party#short_name
This attribute alone causes us a great deal of pain. A party's 'name'
attribute is set to its legal filing entity - e.g. "Libby Schaaf for
Mayor 2014" but we want to display it as just "Libby Schaaf" (hence the
short_name). But this config conflates the candidate configuration
(which I am viewing as us specifically annotating certain committee_id's
with additional attributes that override the ones provided by the state)
with the official data by making it the same object.

This is the first step in breaking apart our configuration from the
source data.

It's unlikely that this commit will work by itself, because I am
separating a bunch of work into commits after-the-fact and there are
some instances of #short_name that I couldn't change in this commit but
will be changed later.