Move candidate configuration into candidate.yml file
This moves all custom candidate configuration into a place that isn't
coupled at all to the Party model. This has the benefit that consumers
of the candidate data can decide whether they want the configuration
only (in case of the campaign page of candidates without data) or if
they want the configuration merged with the Party model's data (in the
case of the campaign page of candidates with data).